+++ example/android/app/src/main/kotlin/com/hwkim1127/simple_native_image_compress_example/MainActivity.kt
@@ -1,6 +1,0 @@
-package com.hwkim1127.simple_native_image_compress_example
-
-import io.flutter.embedding.android.FlutterActivity
-
-class MainActivity: FlutterActivity() {
-}