+++ ios/Classes/simple_native_image_compress.c
@@ -1,6 +1,0 @@
-// Relative import to be able to reuse the C sources.
-// See the comment in ../{projectName}}.podspec for more information.
-// #include "../../src/simple_native_image_compress.c"
-
-
-// This is an empty file to force CocoaPods to create a framework.