+++ lib/simple_native_image_compress.dart
@@ -1,2 +1,5 @@
-export 'src/bridge_definitions.dart' show CompressFormat, FilterType;
-export 'src/simple_native_image_compress.dart';
+library simple_native_image_compress;
+
+export 'src/rust/api/compress.dart';
+export 'src/rust/api/constants.dart';
+export 'src/rust/frb_generated.dart' show NativeImageCompress;